fix: create user, fg

+++ fg/views/admin.py
@@ -1,5 +1,6 @@
 from fg.models import FG
-from fg.serializers import (FGSerializer, FGFileUploadSerializer)
+from fg.serializers import (FGSerializer, FGFileUploadSerializer, CreateFGSerializer)
+from freshman.models import Freshman
 from rest_framework.views import APIView
 from rest_framework.parsers import MultiPartParser
 from rest_framework.response import Response
@@ -10,6 +11,9 @@ from openpyxl import load_workbook
 
 class FGAPI(APIView):
     def get(self, request):
+        if not request.user.is_admin:
+            return Response({"error": True, "data": "Admin role required"})
+        print(request.user.id)
         fg_id = request.GET.get("id")
         error = False
         if fg_id:
@@ -20,17 +24,44 @@ class FGAPI(APIView):
                 data = "FG does not exist"
                 error = True
         else:
-            queryset = FG.objects.all().order_by("-admin")
+            queryset = FG.objects.all().order_by("-is_admin")
             data = FGSerializer(queryset, many=True).data
         return Response({"error": error, "data": data})
 
     def delete(self, request):
+        if not request.user.is_admin:
+            return Response({"error": True, "data": "Admin role required"})
         fg_id = request.GET.get("id")
         if fg_id:
             FG.objects.filter(id=fg_id).delete()
         else:
             FG.objects.all().delete()
         return Response({})
+
+    def post(self, request):
+        if not request.user.is_admin:
+            return Response({"error": True, "data": "Admin role required"})
+        data = request.data
+        serializer = CreateFGSerializer(data=request.data)
+        serializer.is_valid(raise_exception=True)
+        if data["is_admin"] == True:
+            fg = FG.objects.create_super_user(name=data["name"], 
+                            student_id=data["student_id"]
+                            )
+        else:
+            fg = FG.objects.create_user(name=data["name"], 
+                            student_id=data["student_id"]
+                            )
+
+        # try:
+        #     with transaction.atomic():
+        #         Freshman.objects.bulk_create()
+        # except IntegrityError as e:
+        #     return self.Response({"data": str(e).split("\n")[1]})
+
+        data = FGSerializer(fg).data
+        
+        return Response({"error": False, "data": data})
     
 
 class FGFileUploadAPI(APIView):
@@ -38,6 +69,8 @@ class FGFileUploadAPI(APIView):
 
     #TODO: 예외처리
     def post(self, request):
+        if not request.user.is_admin:
+            return Response({"error": True, "data": "Admin role required"})
         serializer = FGFileUploadSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         file = serializer.validated_data["file"]
@@ -55,16 +88,16 @@ class FGFileUploadAPI(APIView):
                     row_value.append(cell.value)
                 fg = FG.objects.create(name=row_value[0], 
                                     student_id=row_value[1],
-                                    admin=row_value[2])
+                                    is_admin=row_value[2])
                 fg_info.append(fg)
             #header 제거
             else:
                 index = 1
-        try:
-            with transaction.atomic():
-                Freshman.objects.bulk_create(freshman_list)
-        except IntegrityError as e:
-            return self.Response({"data": str(e).split("\n")[1]})
+        # try:
+        #     with transaction.atomic():
+        #         Freshman.objects.bulk_create()
+        # except IntegrityError as e:
+        #     return self.Response({"data": str(e).split("\n")[1]})
 
         data = FGSerializer(fg_info, many=True).data
         return Response({"error": error, "data": data})
